cleanup pkg/monitoring
